handle no elements with <title> properly
threw errors when there were no elements <title> in the .rss link. This was normally because there had been no issues reported on that service in the zone

To handle this i split the curl and xmlstarlet separately. I also then counted the amount of elements with title to see if we had any previous issues.