Update check-windows-local-locked-users.ps1
I made my search through event logs more specific. I also only look back now in the last 3 hours for information.

I also now do another call looking for the 4740 which is the locked out, i use this so i can find the source of the lockout (this will be the last computer which makes the account lock).