Update check-ssl-cert-expire.py
added in logic to check that the certificate common name is what we expect it to be. Your webserver you connect to is in domain and the certificate common name is in cert_resp. This is mandatory for you to put in there for this to work

I will eventually update this to handle checking SANs match what you expect.